Added devise gem for user authentication. Fixed bug where user could delete or edit another user's question.

+++ app/controllers/questions_controller.rb
@@ -1,5 +1,7 @@
 class QuestionsController < ApplicationController
 
+  before_filter :authenticate_user!, except: [:index, :show] #does not allow user to modify anything unless logged in.
+
   def index
     @questions = Question.all
   end
@@ -9,12 +11,16 @@ class QuestionsController < ApplicationController
   end
 
   def create
+
     @question = Question.new(question_params)
+    @question.user_id = current_user.id
+
     if @question.save
       redirect_to question_path(@question[:id])
     else
       render :new
     end
+    
   end
 
   def edit
@@ -42,17 +48,23 @@ class QuestionsController < ApplicationController
   def destroy
 
     @question = Question.find(params[:id])
-    @answers = Answer.where question_id: params[:id]
 
-    @answers.each do |answer|
-      answer.destroy
+    if @question.user_id != current_user.id
+      flash[:notice] = "You must be logged in as the question creator to delete this question."
+      redirect_to questions_path(@question)
+    else
+      @answers = Answer.where question_id: params[:id]
+
+      @answers.each do |answer|
+        answer.destroy
+
+      end
+      @question.destroy
+
+      flash[:notice] = "Question deleted."
+      redirect_to questions_path
+
     end
-
-    @question.destroy
-
-
-    flash[:notice] = "Question deleted."
-    redirect_to questions_path
   end
 
   private
